rewrite fetchFromAPI, move funcs into helper dir
