+++ src/fetch/fetchWeatherData.js
@@ -1,7 +1,9 @@
-const https = require('https'),
+const fetchFromAPI = require('./helpers/fetchFromAPI.js'),
+      fetchZipCodes = require('./helpers/fetchZipCodes.js'),
+      weatherAPIEndpoint = require('./helpers/weatherAPIEndpoint.js'),
+      saveWeatherData = require('./helpers/saveWeatherData.js'),
       writeStream = require('../streams/actions/writeStream.js'),
       eventMessages = require('../streams/events/eventMessages.js'),
-      db = require('../db/connection.js'),
       streamName = 'WeatherFetch';
 
 const fetchMsgTimeZones = {
@@ -14,10 +16,6 @@ const fetchMsgTimeZones = {
   FetchUSHawaiiWeather: 'us_hawaii'
 }
 
-const weatherAPIUrl = (zipCode) => {
-  return process.env.WEATHER_API_BASE_URL + `?postal_code=${zipCode}&country=US&units=I&key=${process.env.WEATHER_API_KEY}`;
-}
-
 const fetchWeatherData = async msg => {
   if(!fetchMsgTimeZones[msg]) return;
   // Publish start message to Redis stream
@@ -27,80 +25,29 @@ const fetchWeatherData = async msg => {
         zipCodes = await fetchZipCodes(dbTbl);
 
   zipCodes.forEach(zipCode => {
-    const endPoint = weatherAPIUrl(zipCode);
+    const forecastEndPoint = weatherAPIEndpoint('forecast/daily', zipCode),
+          currentTempEndPoint = weatherAPIEndpoint('current', zipCode);
 
-    fetchFromAPI(endPoint, dbTbl, zipCode);
+    fetchFromAPI(forecastEndPoint)
+      .then(forecastData => {
+        return (
+          fetchFromAPI(currentTempEndPoint)
+            .then(currentTempData => {
+              forecastData['current_temp'] = currentTempData['temp'];
+              return forecastData;
+            })
+           .catch(err => console.error(err))
+        )
+      })
+      .then(data => saveWeatherData(dbTbl, zipCode, data))
+      .catch(err => {
+        writeStream(streamName, eventMessages['error'](err));
+        console.log(err);
+      });
   });
 
   // Publish end message to Redis stream
   writeStream(streamName, eventMessages['end']);
-}
-
-const fetchZipCodes = tbl => (
-  db.select('zip_code')
-    .from(tbl)
-    .then(rows => {
-      return rows.map(row => row.zip_code);
-    })
-    .catch(err => console.error('error:', err))
-)
-
-const fetchFromAPI = (url, dbTbl, zipCode) => {
-  https.get(url, res => {
-    const { statusCode } = res,
-          contentType = res.headers['content-type'];
-
-    let error,
-        rawData = '';
-
-    if (statusCode !== 200) {
-      error = new Error('Request Failed.\n' +
-                        `Status Code: ${statusCode}`);
-    } else if (!/^application\/json/.test(contentType)) {
-      error = new Error('Invalid content-type.\n' +
-                        `Expected application/json but received ${contentType}`);
-    }
-
-    if (error) {
-      writeStream(streamName, eventMessages['error'](error.message));
-      console.error(error.message);
-      // consume response data to free up memory
-      res.resume();
-      return;
-    }
-
-    res.setEncoding('utf8');
-    res.on('data', chunk => rawData += chunk);
-    res.on('end', () => {
-      try {
-        const parsedData = JSON.parse(rawData),
-              weatherData = parsedData.data.slice(0,1);
-
-        saveWeatherData(dbTbl, zipCode, ...weatherData);
-      } catch (err) {
-        writeStream(streamName, eventMessages['error'](err.message));
-        console.error(err.message);
-      }
-    });
-  })
-  .on('error', err => {
-    writeStream(streamName, eventMessages['error'](err.message));
-    console.error(`https.get received an error: ${err.message}`);
-  });
-}
-
-const saveWeatherData = (dbTbl, zipCode, data) => {
-  db(dbTbl)
-    .where({ zip_code: zipCode })
-    .update({ data: data, updated_at: db.fn.now() })
-    .then(() => {
-      writeStream(streamName, eventMessages['data'](zipCode));
-      console.log(`${dbTbl} updated with ${zipCode} weather data`);
-    })
-    .catch(err => {
-      writeStream(streamName, eventMessages['error'](err));
-      console.error('error:', err);
-    })
 }
 
 module.exports = fetchWeatherData;
