+++ src/fetch/helpers/fetchZipCodes.js
@@ -1,0 +1,12 @@
+const db = require('../../db/connection.js');
+
+const fetchZipCodes = tbl => (
+  db.select('zip_code')
+    .from(tbl)
+    .then(rows => {
+      return rows.map(row => row.zip_code);
+    })
+    .catch(err => console.error('error:', err))
+)
+
+module.exports = fetchZipCodes;
